refactor(#3038): extract run_bot S3 gateway helpers
