+++ scripts/giant_file_registry.toml
@@ -101,7 +101,6 @@ grandfathered_baseline_paths = [
   "src/services/discord/router/intake_gate.rs",
   "src/services/discord/router/message_handler/headless_turn.rs",
   "src/services/discord/router/message_handler/intake_turn.rs",
-  "src/services/discord/runtime_bootstrap.rs",
   "src/services/discord/session_runtime.rs",
   "src/services/discord/settings.rs",
   "src/services/discord/tmux.rs",
@@ -203,7 +202,6 @@ grandfathered = [
   "src/services/discord/router/intake_gate.rs",
   "src/services/discord/router/message_handler/headless_turn.rs",
   "src/services/discord/router/message_handler/intake_turn.rs",
-  "src/services/discord/runtime_bootstrap.rs",
   "src/services/discord/session_runtime.rs",
   "src/services/discord/tmux.rs",
   "src/services/discord/tmux_watcher.rs",
